pyproject.toml
[tool.poetry]
name = "ETL postgres_to_elasticsearch"
version = "0.0.1"
description = "Отказоустойчивый перенос данных из Postgres в Elasticsearch"
authors = ["Petr Mazurok <petr.mazurok@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
psycopg2-binary = "^2.9.3"
pydantic = "^1.9.0"
requests = "^2.27.1"
loguru = "^0.6.0"
flake8 = "^4.0.1"
python-dotenv = "^0.20.0"
elasticsearch = "^8.1.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
